+++ src/cli/options/debug.js
@@ -1,0 +1,9 @@
+const debug = {
+  alias: 'd',
+  describe: 'debug yargs parameters',
+  type: 'boolean',
+  default: false,
+  hidden: true
+}
+
+export default debug