+++ src/cli/options/info.js
@@ -1,0 +1,7 @@
+const info = {
+  describe: 'show information and versions',
+  type: 'boolean',
+  conflicts: ['pin', 'use', 'test']
+}
+
+export default info