+++ src/cli/options/use.js
@@ -1,0 +1,10 @@
+import { availablePackages } from '../../packages/list.js'
+
+const use = {
+  alias: 'u',
+  describe: 'use a package manager',
+  choices: availablePackages(),
+  conflicts: ['pin']
+}
+
+export default use