+++ src/flags/test.js
@@ -1,6 +1,6 @@
 import { showCommand } from '../helpers/cmd.js'
 
-export const testCommand = async (pkg) => {
-  await showCommand(pkg)
+export const testCommand = ({ cmd, args, config }) => {
+  showCommand({ cmd, args, config })
   process.exit(0)
 }