+++ src/packages/managers/pnpm.js
@@ -3,7 +3,9 @@ const pnpm = {
   color: '#f7ad24',
   url: 'https://pnpm.io/',
   version: '',
-  lockFile: 'pnpm-lock.yaml'
+  lockFile: 'pnpm-lock.yaml',
+  cmds: {},
+  args: {}
 }
 
 export default pnpm